requirements.txt
requests
python-dotenv
langchain
langchain-community
openai
